zoom to composable
